Renamed wagtailapi module to api
