+++ wagtail/contrib/api/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'wagtail.contrib.api.apps.WagtailAPIAppConfig'
