+++ wagtail/tests/urls.py
@@ -6,7 +6,7 @@ from wagtail.wagtaildocs import urls as wagtaildocs_urls
 from wagtail.wagtailimages import urls as wagtailimages_urls
 from wagtail.wagtailsearch import urls as wagtailsearch_urls
 from wagtail.contrib.wagtailsitemaps.views import sitemap
-from wagtail.contrib.wagtailapi import urls as wagtailapi_urls
+from wagtail.contrib.api import urls as wagtailapi_urls
 
 
 urlpatterns = [
